+++ src/include/file.hpp
@@ -47,6 +47,8 @@ int64_t getFileAgeInDays(const std::string &path);
 
 size_t getMemUsage();
 
+bool isChildProcessAlive(int64_t pid);
+
 bool hasGit();
 
 class FolderLock {